+++ cdisc_rules_engine/services/reporting/base_report.py
@@ -1,7 +1,9 @@
-from abc import abstractmethod, ABC
+from abc import ABC, abstractmethod
 from io import IOBase
-from typing import List, Union, Optional
+from typing import List, Optional, Union
+
 from openpyxl import Workbook
+
 from cdisc_rules_engine.enums.execution_status import ExecutionStatus
 from cdisc_rules_engine.models.rule_validation_result import RuleValidationResult
 from cdisc_rules_engine.models.validation_args import Validation_args
