+++ cdisc_rules_engine/services/reporting/excel_report.py
@@ -1,17 +1,18 @@
 import logging
 from datetime import datetime
-from typing import List, BinaryIO, Optional
+from typing import BinaryIO, List, Optional
 
 from openpyxl import Workbook
 
 from cdisc_rules_engine.models.rule_validation_result import RuleValidationResult
 from cdisc_rules_engine.models.validation_args import Validation_args
-from cdisc_rules_engine.utilities.base_report import BaseReport
-from cdisc_rules_engine.utilities.excel_writer import (
+
+from .base_report import BaseReport
+from .excel_writer import (
     excel_open_workbook,
     excel_update_worksheet,
+    excel_workbook_to_stream,
 )
-from cdisc_rules_engine.utilities.excel_writer import excel_workbook_to_stream
 
 
 class ExcelReport(BaseReport):
