--- tests/resources/dictionaries/whodrug/DDA.txt
@@ -1,5 +1,0 @@
-000001010016C02AB  111*
-000001030014C02AB  074 
-000001040013C02AB  122 
-000002010015R03AB  874*
-000002010015R03CB  874*
